fix(c): clear the cppcheck findings and make the gate blocking
cppcheck was advisory and exiting non-zero, so real C findings could not
stop a merge. 51 findings triaged.

Real defects fixed:
- Four 32-bit multiplies widened only after the fact. umem_size and the
  fill-queue addr both computed num_frames * frame_size in 32 bits before
  assigning to uint64_t, which wraps on a large UMEM.
- Dead inner_etype in reflect_packet_ipv6, assigned twice and then given a
  (void) cast to silence the compiler rather than removed.
- const correctness on select_platform, std_sizes, ip6h and two cfg
  pointers; elapsed scoped to the branch that uses it.

-D__linux__ --platform=unix64 is a correctness fix: -U__APPLE__ removed the
Apple branch but never selected a platform, so cppcheck analysed fallback
stubs and reported their dead branches as defects. That was 5 of 9
knownConditionTrueFalse.

Four suppressions, each justified in the workflow. unusedStructMember is a
cross-TU blind spot -- all 25 were false, disproved by grep.
knownConditionTrueFalse's four survivors are RFC 768's checksum-zero branch
(x2), which it would be a protocol bug to remove, and a static debug
rate-limiter (x2). The const-parameter ones are public API and vtable
signatures, an API change rather than a lint fix.

unreadVariable and variableScope are deliberately left enabled: the former
is what found the dead inner_etype.

Verified on Ubuntu: cppcheck exits 0, clang-tidy still exits 0, and the C
tests pass 9/9.

Closes #858

+++ src/dataplane/common/y1564.c
@@ -123,7 +123,7 @@ static uint64_t __attribute__((unused)) rate_to_pps(double rate_mbps, uint32_t f
 {
     /* Wire size includes preamble (8) + IFG (12) = 20 bytes */
     uint32_t wire_size       = frame_size + 20;
-    uint64_t bits_per_packet = wire_size * 8;
+    uint64_t bits_per_packet = (uint64_t)wire_size * 8;
     uint64_t bits_per_sec    = (uint64_t)(rate_mbps * 1e6);
     return bits_per_sec / bits_per_packet;
 }
